make toast reusable
