+++ frontend/src/components/assessments/EditStatusButtons/DeleteAssessmentButton.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 import { useMutation } from "@apollo/client";
-import { useToast } from "@chakra-ui/react";
 
 import { DELETE_TEST } from "../../../APIClients/mutations/TestMutations";
 import GET_ALL_TESTS from "../../../APIClients/queries/TestQueries";
+import Toast from "../../common/Toast";
 import EditStatusButton from "../EditStatusButton";
 
 interface DeleteAssessmentButtonProps {
@@ -19,35 +19,20 @@ const DeleteAssessmentButton = ({
     refetchQueries: [{ query: GET_ALL_TESTS }],
   });
 
-  const toast = useToast();
-  const showSuccessToast = (message: string) => {
-    toast({
-      description: message,
-      variant: "successToast",
-      status: "success",
-      duration: 9000,
-      isClosable: true,
-      position: "top",
-    });
-  };
-
-  const showErrorToast = (message: string) => {
-    toast({
-      description: message,
-      variant: "errorToast",
-      status: "error",
-      duration: 9000,
-      isClosable: true,
-      position: "top",
-    });
-  };
+  const { showToast } = Toast();
 
   const onDeleteAssessment = async () => {
     await deleteAssessment({ variables: { id: assessmentId } });
     if (error) {
-      showErrorToast("Assessment failed to delete. Please try again.");
+      showToast({
+        message: "Assessment failed to delete. Please try again.",
+        status: "error",
+      });
     } else {
-      showSuccessToast("Assessment deleted.");
+      showToast({
+        message: "Assessment deleted.",
+        status: "success",
+      });
     }
   };
